Backplane/Schematics: Annotate system bus signals
The CPU provides two different sets of high address lines: one for addressing
and another for indication. This makes the schematic confusing, and with so
many unnamed wires it's just hard to read. Add labels for the top part of CPU
slot, change labels on the bottom to conform to Meisaka's notation. Also add a
couple of labels to front panel connector (two LED lines), makes it easier to
verify things.